+++ app/src/main/java/com/example/fitnessfriend/services/Food.java
@@ -14,13 +14,12 @@ public class Food {
     @PrimaryKey
     @NonNull
     String id;
-
     @ColumnInfo(name = "food_id")
     public String foodId;
     @ColumnInfo(name = "label")
     public String label;
     @ColumnInfo(name = "enerc_kcal")
-    public Double ENERC_KCAL;
+    public int ENERC_KCAL;
 
     @ColumnInfo(name = "meal")
     public String meal;
